When merging datasets, allow for not merging images, Convert to RLE, only merge files with matching image directories

+++ coco_assistant/coco_assistant.py
@@ -3,7 +3,7 @@ import json
 import logging
 import os
 import shutil
-
+from pycocotools import mask
 from pycocotools.coco import COCO
 
 from tqdm import tqdm
@@ -15,7 +15,7 @@ from coco_assistant.utils import anchors
 from coco_assistant.utils import det2seg
 
 logging.basicConfig(level=logging.ERROR)
-logging.getLogger().setLevel(logging.WARNING)
+logging.getLogger().setLevel(logging.DEBUG)
 logging.getLogger('parso.python.diff').disabled = True
 
 """
@@ -37,16 +37,20 @@ Expected Directory Structure
 
 
 class COCO_Assistant():
-    def __init__(self, img_dir=None, ann_dir=None):
+    def __init__(self, img_dir=None, ann_dir=None, res_dir=None):
         """
         :param img_dir (str): path to images folder.
         :param ann_dir (str): path to annotations folder.
         """
         self.img_dir = img_dir
         self.ann_dir = ann_dir
-        if os.path.dirname(ann_dir) != os.path.dirname(img_dir):
-            raise AssertionError('Directory not in expected format')
-        self.res_dir = os.path.join(os.path.dirname(ann_dir), 'results')
+        self.res_dir = res_dir
+
+        # if os.path.dirname(ann_dir) != os.path.dirname(img_dir):
+        #     raise AssertionError('Directory not in expected format')
+
+        if self.res_dir is None:
+            self.res_dir = os.path.join(os.path.dirname(ann_dir), 'result')
 
         # Create Results Directory
         if os.path.exists(self.res_dir) is False:
@@ -55,68 +59,80 @@ class COCO_Assistant():
         self.imgfolders = sorted([i for i in os.listdir(self.img_dir) if os.path.isdir(os.path.join(self.img_dir, i)) is True])
         self.jsonfiles = sorted([j for j in os.listdir(ann_dir) if j[-5:] == ".json"])
         self.names = [n[:-5] for n in self.jsonfiles]
+        self.matches = list(set(self.names).intersection(set(self.imgfolders)))
 
-        if self.names != self.imgfolders:
-            raise AssertionError("Image dir and corresponding json file must have the same name")
-
+        logging.debug(f'Names: {self.names}')
+        logging.debug(f'Image folders: {self.imgfolders}')
+        logging.debug(f'Matches: {self.matches}')
+        
         # Note: Add check for confirming these folders only contain .jpg and .json respectively
-        logging.debug("Number of image folders = %s", len(self.imgfolders))
-        logging.debug("Number of annotation files = %s", len(self.jsonfiles))
+        logging.debug("Number of compatible datasets: = %s", len(self.matches))
 
-        if not self.jsonfiles:
+        if not self.names:
             raise AssertionError("Annotation files not passed")
         if not self.imgfolders:
             raise AssertionError("Image folders not passed")
+        if not self.matches:
+            raise AssertionError("Matching Annotation and Image not passed")
 
         self.annfiles = [COCO(os.path.join(ann_dir, i)) for i in self.jsonfiles]
         self.anndict = dict(zip(self.jsonfiles, self.annfiles))
-
         self.ann_anchors = []
 
-    def merge(self):
+    def merge(self, res_dir=None, merge_images=True):
         """
         Function for merging multiple coco datasets
         """
-
-        self.resim_dir = os.path.join(self.res_dir, 'merged', 'images')
-        self.resann_dir = os.path.join(self.res_dir, 'merged', 'annotations')
+        if res_dir:
+            self.res_dir=res_dir
+        self.merge_images = merge_images
+        self.resim_dir = os.path.join(os.path.dirname(self.ann_dir), 'result', 'images')
+        self.resann_dir = os.path.join(os.path.dirname(self.ann_dir), 'result', 'annotations')
 
         # Create directories for merged results and clear the previous ones
         # The exist_ok is for dealing with merged folder
         # TODO: Can be done better
-        if os.path.exists(self.resim_dir) is False:
-            os.makedirs(self.resim_dir, exist_ok=True)
-        else:
-            shutil.rmtree(self.resim_dir)
-            os.makedirs(self.resim_dir, exist_ok=True)
+        
         if os.path.exists(self.resann_dir) is False:
             os.makedirs(self.resann_dir, exist_ok=True)
         else:
             shutil.rmtree(self.resann_dir)
             os.makedirs(self.resann_dir, exist_ok=True)
 
-        print("Merging image dirs")
-        im_dirs = [os.path.join(self.img_dir, folder) for folder in self.imgfolders]
-        imext = [".png", ".jpg"]
+        # Merge images
+        if self.merge_images:
+            if os.path.exists(self.resim_dir) is False:
+                os.makedirs(self.resim_dir, exist_ok=True)
+            else:
+                shutil.rmtree(self.resim_dir)
+                os.makedirs(self.resim_dir, exist_ok=True)
 
-        logging.debug("Merging Image Dirs...")
+            print("Merging image dirs")
+            im_dirs = [os.path.join(self.img_dir, folder) for folder in self.matches]
+            imext = [".png", ".jpg"]
 
-        for imdir in tqdm(im_dirs):
-            ims = [i for i in os.listdir(imdir) if i[-4:].lower() in imext]
-            for im in ims:
-                shutil.copyfile(os.path.join(imdir, im), os.path.join(self.resim_dir, im))
+            logging.debug("Merging Image Dirs...")
 
-        cann = {'images': [],
-                'annotations': [],
-                'info': None,
-                'licenses': None,
-                'categories': None}
+            for imdir in tqdm(im_dirs):
+                ims = [i for i in os.listdir(imdir) if i[-4:].lower() in imext]
+                for im in ims:
+                    shutil.copyfile(os.path.join(imdir, im), os.path.join(self.resim_dir, im))
+
+        else:
+            logging.debug("Not merging images")
+
+        cann = {
+                    'images': [],
+                    'annotations': [],
+                    'info': None,
+                    'licenses': None,
+                    'categories': None
+                }
 
         logging.debug("Merging Annotations...")
 
         dst_ann = os.path.join(self.resann_dir, 'merged.json')
 
-        print("Merging annotations")
         for j in tqdm(self.jsonfiles):
             with open(os.path.join(self.ann_dir, j)) as a:
                 cj = json.load(a)
@@ -126,6 +142,13 @@ class COCO_Assistant():
             # If it is, continue else modify current annotation
             if ind == 0:
                 cann['images'] = cann['images'] + cj['images']
+                
+                #Convert to RLE
+                for i, annotation in enumerate(cj['annotations']):
+                    rle = self.annfiles[ind].annToRLE(annotation)
+                    rle['counts'] = rle['counts'].decode('ascii')
+                    annotation['segmentation'] = rle
+
                 cann['annotations'] = cann['annotations'] + cj['annotations']
                 if 'info' in list(cj.keys()):
                     cann['info'] = cj['info']
@@ -133,11 +156,10 @@ class COCO_Assistant():
                     cann['licenses'] = cj['licenses']
                 cann['categories'] = cj['categories']
 
-                last_imid = cann['images'][-1]['id']
                 last_annid = cann['annotations'][-1]['id']
 
                 # If last imid or last_annid is a str, convert it to int
-                if isinstance(last_imid, str) or isinstance(last_annid, str):
+                if isinstance(last_annid, str):
                     logging.debug("String Ids detected. Converting to int")
                     id_dict = {}
                     # Change image id in images field
@@ -151,23 +173,13 @@ class COCO_Assistant():
                         if isinstance(last_imid, str):
                             im['image_id'] = id_dict[im['image_id']]
 
-                last_imid = cann['images'][-1]['id']
                 last_annid = cann['annotations'][-1]['id']
 
             else:
-
-                id_dict = {}
-                # Change image id in images field
-                for i, im in enumerate(cj['images']):
-                    id_dict[im['id']] = last_imid + i + 1
-                    im['id'] = last_imid + i + 1
-
                 # Change annotation and image ids in annotations field
                 for i, ann in enumerate(cj['annotations']):
                     ann['id'] = last_annid + i + 1
-                    ann['image_id'] = id_dict[ann['image_id']]
 
-                cann['images'] = cann['images'] + cj['images']
                 cann['annotations'] = cann['annotations'] + cj['annotations']
                 if 'info' in list(cj.keys()):
                     cann['info'] = cj['info']
@@ -175,13 +187,12 @@ class COCO_Assistant():
                     cann['licenses'] = cj['licenses']
                 cann['categories'] = cj['categories']
 
-                last_imid = cann['images'][-1]['id']
                 last_annid = cann['annotations'][-1]['id']
 
         with open(dst_ann, 'w') as aw:
             json.dump(cann, aw)
 
-    def remove_cat(self, interactive=True, jc=None, rcats=None):
+    def remove_cat(self, interactive=True, jc=None, rcats=None, res_dir=None):
         """
         Function for removing certain categories.
         In interactive mode, you can input the json and the categories
@@ -194,8 +205,11 @@ class COCO_Assistant():
         :param jc: Json choice
         :param rcats: Categories to be removed
         """
-
-        self.resrm_dir = os.path.join(self.res_dir, 'removal')
+        if res_dir:
+            self.resrm_dir = res_dir
+        else:
+            self.resrm_dir = os.path.join(self.res_dir, 'removal')
+        
         if os.path.exists(self.resrm_dir) is False:
             os.makedirs(self.resrm_dir, exist_ok=True)
         else:
@@ -251,18 +265,18 @@ class COCO_Assistant():
 
         x['categories'] = [i for i in x['categories'] if i not in cats]
         x['annotations'] = [i for i in x['annotations'] if i not in anns]
-
+        print(f"Categories Kept: {x['categories']}")
         with open(os.path.join(self.resrm_dir, self.jc), 'w') as oa:
             json.dump(x, oa)
 
-    def ann_stats(self, stat, arearng, show_count=False, save=False):
+    def ann_stats(self, stat, arearng=None, show_count=False, save=False):
         """
         Function for displaying statistics.
         """
         if stat == "area":
-            stats.pi_area_split(self.annfiles, self.names, areaRng=arearng, save=save)
+            stats.pi_area_split(self.annfiles, self.matches, areaRng=arearng, save=save)
         elif stat == "cat":
-            stats.cat_count(self.annfiles, self.names, show_count=show_count, save=save)
+            stats.cat_count(self.annfiles, self.matches, show_count=show_count, save=save)
 
     def anchors(self, n, fmt=None, recompute=False):
         """
@@ -275,7 +289,7 @@ class COCO_Assistant():
         if recompute or not self.ann_anchors:
             print("Calculating anchors...")
             a = [anchors.generate_anchors(j, n, fmt) for j in self.annfiles]
-            self.ann_anchors = dict(zip(self.names, a))
+            self.ann_anchors = dict(zip(self.matches, a))
         else:
             print("Loading pre-computed anchors")
             print(self.ann_anchors)
@@ -284,7 +298,7 @@ class COCO_Assistant():
         """
         Function for generating segmentation masks.
         """
-        for ann, name in zip(self.annfiles, self.names):
+        for ann, name in zip(self.annfiles, self.matches):
             output_dir = os.path.join(self.res_dir, 'segmasks', name)
             det2seg.det2seg(ann, output_dir)
 
@@ -295,13 +309,13 @@ class COCO_Assistant():
         :param to: Format to which annotations are to be converted
         """
         print("Choose directory:")
-        print(self.imgfolders)
+        print(self.matches)
 
         dir_choice = input()
 
-        if dir_choice.lower() not in [item.lower() for item in self.imgfolders]:
+        if dir_choice.lower() not in [item.lower() for item in self.matches]:
             raise AssertionError("Choice not in images folder")
-        ind = self.imgfolders.index(dir_choice.lower())
+        ind = self.matches.index(dir_choice.lower())
         ann = self.annfiles[ind]
         img_dir = os.path.join(self.img_dir, dir_choice)
 
@@ -312,20 +326,20 @@ class COCO_Assistant():
         Function for visualising annotations.
         """
         print("Choose directory:")
-        print(self.imgfolders)
+        print(self.matches)
 
         dir_choice = input()
 
-        if dir_choice.lower() not in [item.lower() for item in self.imgfolders]:
+        if dir_choice.lower() not in [item.lower() for item in self.matches]:
             raise AssertionError("Choice not in images folder")
-        ind = self.imgfolders.index(dir_choice.lower())
+        ind = self.matches.index(dir_choice.lower())
         ann = self.annfiles[ind]
         img_dir = os.path.join(self.img_dir, dir_choice)
         cocovis.visualise_all(ann, img_dir)
 
 
 if __name__ == "__main__":
-    p = "/home/ashwin/Desktop/Projects/COCO-Assistant/data/test"
+    p = "./data/test"
     img_dir = os.path.join(p, 'images')
     ann_dir = os.path.join(p, 'annotations')
